add api for loading img
