+++ src/index.html
@@ -34,9 +34,9 @@
     <main> 
 
     <div class="catalogs"></div>
-
-           <div id="movie-display"></div>
-    </main>
+    
+  
+</main>
