+++ src/style.css
@@ -98,8 +98,20 @@ h2 {
   font-weight: 400;
   display: flex;
   margin-left: 47%;
-  margin-top: -16px;
+  margin-top: -19px;
+ 
 }
+
+.count-like {
+  font-family: 'Raleway', sans-serif;
+  font-weight: 400;
+  display: flex;
+  margin-left: 42%;
+  margin-top: -35px;
+  padding-top: 13px;
+  
+}
+
 
 .commentsBtn {
   padding: 5px 37px;
@@ -107,6 +119,7 @@ h2 {
   font-weight: 400;
   font-size: 15px;
   transform: scale(1.05);
+  display: block;
 }
 
 .commentsBtn:hover {
@@ -127,47 +140,65 @@ h2 {
 }
 
 /* ----------popup------ */
-/* 
+
 .movie-display {
     align-items: center;
     margin-left: 50px;
-} */
+}
+#popupFormContainer {
+  display: none;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  padding: 26px;
+  border-radius: 5px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  z-index: 9999;
+  font-size: 14px;
+  width: 30%;
+}
 
-#movie-display {
-    display: none;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    z-index: 9999;
-   
-  }
+#popupFormContainer.open {
+  display: block;
+}
+
+#closeButton {
+  margin-top: 10px;
+}
+
+#closeButton{
+  float: right;
+  font-size: 25px;
+  font-weight: bold;
+  cursor: pointer;
+  
+}
 
 
+/* 
 .fa-xmark {
     position: absolute;
     padding-left: 50%;
     width: 50%;
-}
+} */
 
 .summary {
-    font-family: 'Raleway', sans-serif;
-    font-weight: 400;
-    letter-spacing: -0.0458;
+    /* font-family: 'Raleway', sans-serif;
+    font-weight: 400; */
+    
 }
 
-h4 {
-    font-family: 'Raleway', sans-serif;
-    font-weight: 400;
-    letter-spacing: -0.0458;
+h2 {
+    font-family:  sans-serif;
+    /* letter-spacing: -0.0458; */
 }
 
 h3 {
     font-family: 'Raleway', sans-serif;
     font-weight: 400;
-    letter-spacing: -0.0458;
+   
 }
 /* ---------footer---------- */
 footer p {
