+++ src/template.html
@@ -24,7 +24,7 @@
         <nav>
             <a href="#" class="logo"><span id="logo-A">A</span>frimoviz</a>
             <ul>
-                <li><a href="#">showmovie</a></li>
+                <li><a href="#">showmovie<span id="cardCount"></span></a></li>
                 <li><a href="#">buy movies</a></li>
                 <li><a href="#">subscribe</a></li>
             </ul>
@@ -34,6 +34,60 @@
     <main> 
 
     <div class="catalogs"></div>
+    <!--popup-->
+    <button id="commentsBtn">comment</button>
+    <div id="popupFormContainer">
+      <div id="movie-display">
+        
+          <div class="photo1">
+            <span id="closeButton"><i class="fa-solid fa-xmark"></i></span>
+            
+              <img class="movi-pic" src="../assets/images/blacklist.png" alt="movie phonto">
+              
+          </div>
+  
+          <h2> The Black list</h2>
+          <div class="deatails">
+              <div class="df1">
+                  <h2>Language:English</h2>
+                  <h2>Estimated time:2hrs</h2>
+              </div>
+              <div class="df2">
+                  <h2>Movie genres:Drama</h2>
+                  <h2>Rating:5.5</h2>
+              </div>
+          </div>
+          <h2>Summary:</h2>
+          <p class="summary">
+              The Blacklist is a crime drama involving a former government agent who turned into a high-profile criminal
+              turning himself in to the FBI offering to help catch criminals.
+          </p>
+          <div class="whole-coment">
+              <h3>Add your comments</h3>
+              
+          </div>
+      </div>
+  
+      
+          <form id="popupForm" class="form">
+              <div class="flex-form">
+                  <input type="text" name="name" placeholder="Your Name" class="name" id="name"> <br>
+                  <textarea
+                      name="text-message"
+                      class="type-text message"
+                      cols="30"
+                      rows="10"
+                      placeholder="Your insight"
+                      id="message"
+                      maxlength="500"
+                      required
+                  ></textarea>
+              </div>
+              
+              <button id="commentsBtn" type="submit">Comment</button>
+          </form>
+      </div>
+    </div>
 
        
     </main>
